Main restructure complete and .vscode task config updated so all files build
